+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Fixed rare crash due to calling new scan before the previous onScanEnd callback was complete. 
+- Removed serial printf's during updates as it was causing occasional update issues.
 
 ### Hardware